+++ maven-confluence-reporting-plugin/src/main/java/org/bsc/maven/confluence/plugin/AbstractBaseConfluenceMojo.java
@@ -1,10 +1,5 @@
 package org.bsc.maven.confluence.plugin;
 
-import static java.lang.String.format;
-
-import java.util.Optional;
-import java.util.function.Consumer;
-
 import org.apache.maven.plugin.AbstractMojo;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugins.annotations.Component;
@@ -22,6 +17,11 @@ import org.sonatype.plexus.components.sec.dispatcher.DefaultSecDispatcher;
 import org.sonatype.plexus.components.sec.dispatcher.SecDispatcher;
 import org.sonatype.plexus.components.sec.dispatcher.SecDispatcherException;
 
+import java.util.Optional;
+import java.util.function.Consumer;
+
+import static java.lang.String.format;
+
 
 /**
  *
@@ -30,9 +30,10 @@ import org.sonatype.plexus.components.sec.dispatcher.SecDispatcherException;
 public abstract class AbstractBaseConfluenceMojo extends AbstractMojo {
     /**
      * additional properties pass to template processor
+     * Properties in the form of URI will be loaded and loaded value will be used instead, see processProperties
      */
     @Parameter()
-    private java.util.Map<String, String> properties;    
+    private java.util.Map<String, String> properties;
     /**
      * Confluence api endpoint url
      * 
@@ -196,7 +197,7 @@ public abstract class AbstractBaseConfluenceMojo extends AbstractMojo {
 
     public final java.util.Map<String, String> getProperties() {
         if (null == properties) {
-            properties = new java.util.HashMap<String, String>(5);
+            properties = new java.util.HashMap<>(5);
         }
         return properties;
     }