+++ maven-confluence-reporting-plugin/src/main/java/org/bsc/maven/confluence/plugin/ConfluenceDeployMojo.java
@@ -3,6 +3,25 @@ package org.bsc.maven.confluence.plugin;
 import biz.source_code.miniTemplator.MiniTemplator;
 import biz.source_code.miniTemplator.MiniTemplator.VariableNotDefinedException;
 import com.github.qwazer.mavenplugins.gitlog.CalculateRuleForSinceTagName;
+import static java.lang.String.format;
+import static java.util.concurrent.CompletableFuture.completedFuture;
+import static org.bsc.confluence.model.SitePrinter.print;
+import static org.bsc.confluence.model.SiteProcessor.processPageUri;
+
+import java.io.StringWriter;
+import java.nio.file.Path;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Locale;
+import java.util.Map;
+import java.util.Objects;
+import java.util.Optional;
+import java.util.Set;
+import java.util.concurrent.CompletableFuture;
+
+import org.apache.commons.lang.StringUtils;
 import org.apache.maven.artifact.Artifact;
 import org.apache.maven.artifact.factory.ArtifactFactory;
 import org.apache.maven.artifact.metadata.ArtifactMetadataSource;
@@ -227,6 +246,14 @@ public class ConfluenceDeployMojo extends AbstractConfluenceDeployMojo {
     @Parameter(defaultValue="false")
     private Boolean gitLogGroupByVersions;
 
+    /**
+     * Overrides system locale used for content generation
+     *
+     * @since 6.6
+     */
+    @Parameter
+    private String locale;
+
 
     @Override
     public void execute() throws MojoExecutionException, MojoFailureException {
@@ -246,7 +273,7 @@ public class ConfluenceDeployMojo extends AbstractConfluenceDeployMojo {
 
 		}
 
-        final Locale locale = Locale.getDefault();
+        final Locale parsedLocale = !StringUtils.isEmpty(locale) ? new Locale(locale) : Locale.getDefault();
 
         getLog().info(format("executeReport isSnapshot = [%b] isRemoveSnapshots = [%b]", isSnapshot(), isRemoveSnapshots()));
 
@@ -296,14 +323,14 @@ public class ConfluenceDeployMojo extends AbstractConfluenceDeployMojo {
            // PLUGIN
            /////////////////////////////////////////////////////////////////
             {
-                generatePluginReport(site, locale);
+                generatePluginReport(site, parsedLocale);
             }
             else
            /////////////////////////////////////////////////////////////////
            // PROJECT
            /////////////////////////////////////////////////////////////////
             {
-                generateProjectReport(site, locale);
+                generateProjectReport(site, parsedLocale);
             }
 
         } catch( MojoExecutionException e ) {
@@ -545,9 +572,11 @@ public class ConfluenceDeployMojo extends AbstractConfluenceDeployMojo {
                 final Model.Page homePage,
                 final Locale locale )
     {
-        final java.net.URI uri = site.getHome().getUri();
 
-        return processPageUri( site, homePage, uri, homePage.getTitle(), (err, content) -> {
+        final Site.Page home = site.getHome();
+        final java.net.URI uri = home.getUri();
+
+        return processPageUri( site, home, homePage, uri, homePage.getTitle(), (err, content) -> {
             final CompletableFuture<Model.Page> result = new CompletableFuture<>();
 
             try {
@@ -897,7 +926,7 @@ public class ConfluenceDeployMojo extends AbstractConfluenceDeployMojo {
 
             final String title = getPageTitle();
 
-            return processPageUri(site, homePage, site.getHome().getUri(), getPageTitle(), ( err, content ) -> {
+            return processPageUri(site, site.getHome(), homePage, site.getHome().getUri(), getPageTitle(), ( err, content ) -> {
 
                 final CompletableFuture<Model.Page> result =
                         new CompletableFuture<>();